% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemfunctions.R
\name{computeTau}
\alias{computeTau}
\title{The posterior probability}
\usage{
computeTau(x, p, mu, sigma)
}
\arguments{
\item{x}{A matrix of observations with d rows (number of variables)
and n columns (number of observations).}

\item{p}{A numeric vector of proportions (one for each mixture component).}

\item{mu}{A list of numeric vectors (which are the mean of each mixture component).}

\item{sigma}{A list of covariance matrices (one for each mixture component).}
}
\value{
This function returns a matrix that contains the posterior probabilities
}
\description{
This function calculates the posterior probability that each observation belongs to each of the mixture components.
}
