% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemfunctions.R
\name{gemEstimator}
\alias{gemEstimator}
\title{Estimates the Gaussian mixture parameters using GEM algorithm}
\usage{
gemEstimator(x, Nit, p0, mu0, sigma0, G = list())
}
\arguments{
\item{x}{A matrix of observations with d rows (number of variables) and n columns (number of observations).}

\item{Nit}{An integer that specifies the iterations number. If a negative value is
specified then a stopping rule is used.}

\item{p0}{A numeric vector of initial proportions.}

\item{mu0}{A list of numeric vectors which are the initial mean of each
mixture component.}

\item{sigma0}{A list of covariance matrices (Algorithm initialization)}

\item{G}{A list of K graphs associated with the mixture components. Each
graph is a named list with three items : Cliques list, Sperators list and
Sperators multiplicities. If an empty list is provided then the classical
EM algorithm is used.}
}
\value{
This function returns a named list containing the following results:
'p': A vector of estimated proportions.
'mu': A list of estimated mean vectors.
'sigma' : A list of estimated covarianece matrices which takes into
account the provided graph structures.
'tau' : An n by K matrix that contains the posterior probabilities that each observation belongs to each mixture component.
'NbIterations' ; The number of performed iterations.
}
\description{
This function estimates the Gaussian mixture parameters using the GEM
algorithm. The mixture components number is supposed to be known.
This function uses an initial parameters guess and a set of associated
graphs to iteratively estimate the parameters.
}
