% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemfunctions.R
\name{graphMatrixAssoc}
\alias{graphMatrixAssoc}
\title{Measure the association degree between a graph and a covariance Matrix}
\usage{
graphMatrixAssoc(A, C)
}
\arguments{
\item{A}{A covariance Matrix.}

\item{C}{A list of numeric vectors which specify the graph's maximal cliques}
}
\value{
This function returns a numeric value which measures the association degree between the covariance matrix and the graph.
}
\description{
This function computes the association degree between a covariabce matrix
and a graph. The computed metric relies only on the correspondance between
the zeros in the inverted covariance matix and the set of conditional independencies.
}
\examples{
A <- matrix(0,5,5)
diag(A) <- 1:5
diag(A[-1,-5]) <- 1:4
diag(A[-5,-1]) <- 1:4
cliques <- list(c(1,2),c(2,3), c(3,4),c(4,5))
d1 <- graphMatrixAssoc(A,C)

}
