% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemfunctions.R
\name{normTheta}
\alias{normTheta}
\title{Computes the norm of a set of parameters.}
\usage{
normTheta(p, mu, sigma)
}
\arguments{
\item{p}{A numeric vector of size K.}

\item{mu}{A list of K numeric vectors of size d.}

\item{sigma}{A list of K square matrices of size dxd}
}
\value{
This function returns a numeric value which is the norm of the provided
parameters.
}
\description{
This function computes the norm of the Gaussian mixture parameters set as the
sum of the respective norms of each element.
}
